feat(lights): :sparkles: update LightsCard layout and implement LightsCardViewHolder for improved UI and functionality
