+++ app/src/main/res/layout/item_lights_card.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!-- item_lights_card.xml -->
+<?xml version="1.0" encoding="utf-8"?>
+<!-- item_lights_card.xml -->
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/cardView"
@@ -56,52 +57,14 @@
                 android:textColor="@color/slateGrayDark"
                 android:textSize="16sp"
                 android:textStyle="bold" />
-
-            <!--            <TextView-->
-            <!--                android:id="@+id/tvSubtitle"-->
-            <!--                android:layout_width="wrap_content"-->
-            <!--                android:layout_height="wrap_content"-->
-            <!--                android:layout_marginTop="2dp"-->
-            <!--                android:ellipsize="end"-->
-            <!--                android:maxLines="1"-->
-            <!--                android:text="1 Task"-->
-            <!--                android:textColor="@color/textGrayDark"-->
-            <!--                android:textSize="12sp" />-->
         </LinearLayout>
 
-        <!-- Toggle Switch -->
-        <androidx.cardview.widget.CardView
-            android:id="@+id/switch_container"
-            android:layout_width="80dp"
-            android:layout_height="32dp"
-            android:backgroundTint="@color/brightBlue"
-            app:cardCornerRadius="16dp"
-            app:cardElevation="2dp">
-
-            <!-- Texto ON/OFF dentro del switch -->
-            <TextView
-                android:id="@+id/tv_switch_text"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center_vertical"
-                android:layout_marginStart="8dp"
-                android:text="ON"
-                android:textColor="@color/white"
-                android:textSize="10sp"
-                android:textStyle="bold" />
-
-            <!-- Thumb del switch -->
-            <androidx.cardview.widget.CardView
-                android:id="@+id/switch_thumb"
-                android:layout_width="24dp"
-                android:layout_height="24dp"
-                android:layout_gravity="start|center_vertical"
-                android:layout_marginStart="4dp"
-                android:backgroundTint="@color/white"
-                app:cardCornerRadius="12dp"
-                app:cardElevation="4dp" />
-
-        </androidx.cardview.widget.CardView>
+        <!--switch -->
+        <com.google.android.material.switchmaterial.SwitchMaterial
+            android:id="@+id/material_switch"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:checked="true" />
 
     </LinearLayout>
